backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncmy
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
